Moving generic api's to library and optimising intent modules
